Move yum_util.py into utils/
*OSBS-8605

Also change files which import yum_util.py
